fix when $file_name -eq $name

+++ exe.ps1
@@ -30,12 +30,14 @@ $ErrorActionPreference = 'Stop'
 try {
 	# absolute path is required, because current directory is not the same as the script directory
 	$file_name, $etag = ~/.gpm/scripts/lib/gh_dl.ps1 -repo $repo -ScriptBlock { $_.name -match $pattern }
+
+	if ($name -ne $file_name) {
+		Remove-Item $name -ErrorAction SilentlyContinue
+		Rename-Item -LiteralPath $file_name $name -Force
+	}
 } catch {
 	[Console]::Error.WriteLine($_.Exception.Message)
 	exit 1
 }
-
-Remove-Item $name -ErrorAction SilentlyContinue
-Rename-Item -LiteralPath $file_name $name -Force
 
 $etag
